fixed #36 and rearranged/removed board.py
