+++ enviro/boards/grow.py
@@ -2,7 +2,7 @@ import time
 from breakout_bme280 import BreakoutBME280
 from breakout_ltr559 import BreakoutLTR559
 from machine import Pin, PWM
-from enviro.board import i2c
+from enviro import i2c
 
 bme280 = BreakoutBME280(i2c, 0x77)
 ltr559 = BreakoutLTR559(i2c)
