+++ main.py
@@ -43,7 +43,6 @@ if user_requested_provisioning or enviro.needs_provisioning():
 # import config now that we know provisioning isn't needed
 import config
 
-
 # if the clock isn't set then we need to fetch the time from an NTP
 # server. this requires connecting to WiFi
 if not enviro.clock_set():
@@ -52,7 +51,7 @@ if not enviro.clock_set():
     # if we failed to synchronise the clock then turn on the warning
     # led and go back to sleep for another cycle
     logging.error("! failed to synchronise clock")
-    warn_led(WARN_LED_BLINK)
+    enviro.warn_led(enviro.WARN_LED_BLINK)
     enviro.sleep(config.reading_frequency)
 
 
@@ -61,7 +60,7 @@ if enviro.low_disk_space():
   # means that cached results are not getting uploaded and cleared so
   # warn the user and go back to sleep
   logging.error("! low disk space")
-  warn_led(WARN_LED_BLINK)
+  enviro.warn_led(enviro.WARN_LED_BLINK)
   enviro.sleep(config.reading_frequency)
 
 filesystem_stats = os.statvfs(".")
